refactor: rename modules to fit and synthesis
