+++ ROADMAP.md
@@ -1,7 +1,7 @@
 # Nemora Readiness Roadmap
 
 This roadmap tracks Nemora’s expansion from a distribution-fitting prototype into a modular
-meta-package. It is intentionally aggressive—distfit is sprinting to alpha—while other modules
+meta-package. It is intentionally aggressive—fit is sprinting to alpha—while other modules
 will land in sequenced phases. The plan is updated alongside daily work; if something feels out of
 date, check `notes/nemora_modular_reorg_plan.md` for the latest detail.
 
@@ -9,36 +9,36 @@ date, check `notes/nemora_modular_reorg_plan.md` for the latest detail.
 - Repository scaffolding, licensing, CI harness.
 - Initial documentation + contribution standards.
 - Core refactor: renamed project to Nemora, bootstrapped `nemora.core`, centralised distribution
-  registry, migrated fitting stack into `nemora.distfit`, added compatibility shims.
+  registry, migrated fitting stack into `nemora.fit`, added compatibility shims.
 
 ## Phase 1 — Distribution Fitting Alpha 🚧 (in progress, target: next 1–2 days)
-- [x] Finalise `nemora.distfit` alpha API (grouped EM, mixtures, goodness-of-fit).
-- [x] Expand distfit unit tests (fixtures, CLI regressions, coverage gating).
+- [x] Finalise `nemora.fit` alpha API (grouped EM, mixtures, goodness-of-fit).
+- [x] Expand fit unit tests (fixtures, CLI regressions, coverage gating).
 - [x] Publish module overview + API reference; update README/CLI help.
 - [x] Verify notebooks/examples reference the new namespace.
-- [x] Cut changelog entry announcing distfit alpha and note breaking import changes.
+- [x] Cut changelog entry announcing fit alpha and note breaking import changes.
 
-## Phase 2 — Core Module Expansion (sequenced after distfit alpha)
+## Phase 2 — Core Module Expansion (sequenced after fit alpha)
 - `nemora.distributions`
   - [x] Document extension points, add user-facing registry helpers.
   - [ ] Move remaining distribution metadata (bounds, defaults) from ad-hoc code.
 - `nemora.sampling`
   - [x] Implement PDF→CDF inversion (analytic + numeric).
   - [x] Provide bootstrap / Monte Carlo sampling utilities & tests.
-  - [x] Integrate mixture helpers with distfit outputs.
+  - [x] Integrate mixture helpers with fit outputs.
 - `nemora.ingest`
   - [x] Design abstraction for raw inventory sources (`DatasetSource`, `TransformPipeline`).
   - [x] Port existing scripts (HPS dataset prep) into pipelines.
   - [x] Add CLI helpers for fetching / transforming reference datasets.
   - [x] Publish ingest how-to updates covering FAIB/FIA workflows.
   - [x] Add regression coverage for FAIB manifest + pipeline orchestration.
-- `nemora.synthforest`
+- `nemora.synthesis`
   - [ ] Define landscape/stem data models.
   - [ ] Implement stand attribute sampling, stem population generation.
   - [ ] Provide export & visualisation helpers (GeoJSON, rasters).
 - `nemora.simulations`
   - [ ] Create interfaces for plot-based and remote-sensing simulations.
-  - [ ] Integrate with synthforest outputs and sampling utilities.
+  - [ ] Integrate with synthesis outputs and sampling utilities.
   - [ ] Model measurement noise / detection bias; deliver CLI workflows.
 
 ## Phase 3 — Quality & Release Readiness
@@ -57,19 +57,23 @@ date, check `notes/nemora_modular_reorg_plan.md` for the latest detail.
 
 ## Detailed Next Steps Notes
 1. **Distribution registry hardening**
-   - [x] Inventory bounds/defaults/extras scattered across distfit, sampling, and ingest helpers; relocate the metadata into `src/nemora/distributions` and surface it through helper APIs.
+   - [x] Inventory bounds/defaults/extras scattered across fit, sampling, and ingest helpers; relocate the metadata into `src/nemora/distributions` and surface it through helper APIs.
    - [x] Add a CLI/`nemora registry` subcommand (and Python helper) that prints the registered bounds/defaults/extras so contributors can inspect metadata without diving into code.
    - [x] Expand regression tests covering the new helper/CLI output (including plugin registration coverage) and ensure failure modes stay readable.
    - [x] Publish an updated registry reference (docs + README) connecting the Python helpers, CLI inspection, YAML configs, and entry-point plugins.
 2. **Sampling module adoption**
-   - [x] Wire `BootstrapResult` into upcoming synthforest/simulation plans and document how downstream modules consume its metadata (bootstrap payload helper + docs).
+   - [x] Wire `BootstrapResult` into upcoming synthesis/simulation plans and document how downstream modules consume its metadata (bootstrap payload helper + docs).
    - [x] Add numerical accuracy tests that exercise the trapezoid/Simpson/quad integration modes against SciPy references.
    - [x] Extend docs/examples so sampling workflows demonstrate ingest-produced Parquet manifests and the configurable numeric integration settings.
-   - [x] Prototype synthforest/simulation adapters that consume the new helper (CLI wiring + regression tests).
+   - [x] Prototype synthesis/simulation adapters that consume the new helper (CLI wiring + regression tests).
 3. **Ingest monitoring & benchmarking**
    - [x] Capture `nemora ingest-benchmark` runtime stats (CLI + nightly workflow) and surface the trend in docs or CHANGE_LOG for visibility.
    - [x] Document the nightly FAIB/FIA workflow rerun + notification process in `CONTRIBUTING.md` (plus the new benchmark summary/threshold automation) so contributors can verify the job locally.
    - [x] Evaluate whether manifest Parquet adoption should become the default artifact once benchmarks confirm no downstream regressions (CLI now writes CSV + Parquet by default with `--no-parquet` opt-out).
+4. **Module naming alignment**
+   - [ ] Rename `nemora.fit` to `nemora.fit` while preserving a compatibility shim + CLI aliases so downstream docs/tests stay stable during the transition.
+   - [ ] Rename `nemora.synthforest` to `nemora.synthesis`, folding forest/stand/tree-level generators under the clearer namespace and updating planning docs/notes accordingly.
+   - [ ] Announce the breaking rename in README/CHANGE_LOG once the shims/tests/CLI wiring are proven on a feature branch.
 
 ## Backlog & Ideas
 - [ ] Investigate GPU acceleration for large tally batches.
