--- docs/api/distfit.md
@@ -1,7 +1,0 @@
-# `nemora.distfit` module
-
-```{automodule} nemora.distfit
-:members:
-:undoc-members:
-:show-inheritance:
-```
