+++ docs/api/fit.md
@@ -1,0 +1,7 @@
+# `nemora.fit` module
+
+```{automodule} nemora.fit
+:members:
+:undoc-members:
+:show-inheritance:
+```
