+++ docs/reference/distributions.md
@@ -108,4 +108,4 @@ starting values that can be overridden via `FitConfig`.
   - Log-logistic distribution with equal shape parameters.
 ```
 
-.. todo:: Update this section once the nemora.ingest / sampling / synthforest modules land to reflect the broader workflow.
+.. todo:: Update this section once the nemora.ingest / sampling / synthesis modules land to reflect the broader workflow.
