+++ docs/reference/index.md
@@ -4,5 +4,5 @@
 :maxdepth: 1
 
 distributions
-distfit
+fit
 ```
