+++ src/nemora/fit/grouped.py
