+++ src/nemora/synthforest/__init__.py
@@ -1,11 +1,16 @@
-"""Helper utilities for synthforest consumers."""
+"""Compatibility shim for the renamed :mod:`nemora.synthesis` package."""
 
 from __future__ import annotations
 
-from .helpers import BootstrapPayload, bootstrap_payload, bootstrap_to_dataframe
+import warnings
 
-__all__ = [
-    "BootstrapPayload",
-    "bootstrap_payload",
-    "bootstrap_to_dataframe",
-]
+from ..synthesis import *  # noqa: F401,F403
+from ..synthesis import __all__ as _synthesis_all
+
+warnings.warn(
+    "`nemora.synthforest` has been renamed to `nemora.synthesis`; import from the new namespace.",
+    DeprecationWarning,
+    stacklevel=2,
+)
+
+__all__ = _synthesis_all
