+++ src/nemora/workflows/censoring.py
@@ -7,7 +7,7 @@ from collections.abc import Iterable, Mapping
 import numpy as np
 
 from ..core import FitResult, InventorySpec
-from ..distfit import FitConfig, default_fit_config, fit_inventory, fit_with_lmfit
+from ..fit import FitConfig, default_fit_config, fit_inventory, fit_with_lmfit
 
 
 def fit_censored_inventory(
