+++ src/nemora/workflows/hps.py
@@ -7,7 +7,7 @@ from collections.abc import Iterable, Mapping
 import numpy as np
 
 from ..core import FitResult, InventorySpec
-from ..distfit import FitConfig, default_fit_config, fit_inventory
+from ..fit import FitConfig, default_fit_config, fit_inventory
 from ..weighting import hps_compression_factor, hps_expansion_factor
