+++ tests/test_grouped_fixtures.py
@@ -7,7 +7,7 @@ import pandas as pd
 import pytest
 
 from nemora.core import InventorySpec
-from nemora.distfit import fit_inventory
+from nemora.fit import fit_inventory
 from nemora.weighting import hps_expansion_factor
 from nemora.workflows.hps import fit_hps_inventory
